+++ addon/components/bootstrap-button.js
@@ -14,20 +14,11 @@ export default Ember.Component.extend(PropTypeMixin, {
     disabled: PropTypes.bool,
   },
 
-  /* private attrs */
-  _isRunning: false,
-
-  disabledButton: Ember.computed.or('_isRunning', 'disabled'),
+  disabledButton: Ember.computed.or('asyncTask.isRunning', 'disabled'),
 
   asyncTask: task(function * (asyncTask) {
-    this.set('_isRunning', true);
-
-    try {
       return yield asyncTask();
-    } finally {
-      this.set('_isRunning', false);
-    }
-  }),
+  }).drop(),
 
   click() {
     return this.get('asyncTask').perform(this.get('action')).catch(e => {